+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/automatiskVurdering/VerdikjedeTest.kt
@@ -1,31 +1,48 @@
 package no.nav.familie.ba.sak.kjerne.automatiskVurdering
 
+import io.mockk.clearAllMocks
 import io.mockk.every
 import no.nav.familie.ba.sak.common.DbContainerInitializer
-import no.nav.familie.ba.sak.common.randomFnr
+import no.nav.familie.ba.sak.config.ClientMocks
+import no.nav.familie.ba.sak.config.ClientMocks.Companion.initEuKodeverk
+import no.nav.familie.ba.sak.config.FeatureToggleConfig
+import no.nav.familie.ba.sak.config.FeatureToggleService
 import no.nav.familie.ba.sak.config.e2e.DatabaseCleanupService
-import no.nav.familie.ba.sak.integrasjoner.infotrygd.InfotrygdService
+import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.IntegrasjonClient
+import no.nav.familie.ba.sak.integrasjoner.infotrygd.domene.InfotrygdFødselhendelsesFeedTaskDto
 import no.nav.familie.ba.sak.integrasjoner.pdl.PdlRestClient
 import no.nav.familie.ba.sak.integrasjoner.pdl.PersonopplysningerService
-import no.nav.familie.ba.sak.kjerne.automatiskvurdering.FagsystemRegelVurdering
-import no.nav.familie.ba.sak.kjerne.automatiskvurdering.VelgFagSystemService
-import no.nav.familie.ba.sak.kjerne.behandling.NyBehandlingHendelse
+import no.nav.familie.ba.sak.integrasjoner.pdl.VergeResponse
+import no.nav.familie.ba.sak.integrasjoner.pdl.internal.DødsfallData
+import no.nav.familie.ba.sak.kjerne.automatiskvurdering.FiltreringsreglerResultat
+import no.nav.familie.ba.sak.kjerne.behandling.BehandlingService
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingResultat
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakService
+import no.nav.familie.ba.sak.kjerne.fagsak.FagsakStatus
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlagRepository
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.domene.PersonIdent
 import no.nav.familie.ba.sak.kjerne.steg.StegService
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.BeforeAll
+import no.nav.familie.ba.sak.kjerne.steg.StegType
+import no.nav.familie.ba.sak.kjerne.vedtak.VedtakService
+import no.nav.familie.ba.sak.task.BehandleFødselshendelseTask
+import no.nav.familie.ba.sak.task.dto.OpprettOppgaveTaskDTO
+import no.nav.familie.kontrakter.felles.objectMapper
+import no.nav.familie.kontrakter.felles.personopplysning.Ident
+import no.nav.familie.prosessering.domene.TaskRepository
+import org.junit.Assert.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.context.junit.jupiter.SpringExtension
+import java.time.LocalDate
 
 @SpringBootTest(properties = ["FAMILIE_FAMILIE_TILBAKE_API_URL=http://localhost:28085/api"])
 @ExtendWith(SpringExtension::class)
@@ -39,40 +56,224 @@ import org.springframework.test.context.junit.jupiter.SpringExtension
         "mock-infotrygd-barnetrygd",
 )
 @Tag("integration")
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class Integrasjonstest(
+class VerdikjedeTest(
         @Autowired val stegService: StegService,
         @Autowired val personopplysningerService: PersonopplysningerService,
         @Autowired val persongrunnlagService: PersongrunnlagService,
         @Autowired val personopplysningGrunnlagRepository: PersonopplysningGrunnlagRepository,
         @Autowired val pdlRestClient: PdlRestClient,
-        @Autowired val velgFagSystemService: VelgFagSystemService,
         @Autowired val fagSakService: FagsakService,
-        @Autowired val infotrygdService: InfotrygdService,
-        @Autowired val databaseCleanupService: DatabaseCleanupService
+        @Autowired val taskRepository: TaskRepository,
+        @Autowired val behandleFødselshendelseTask: BehandleFødselshendelseTask,
+        @Autowired val behandlingService: BehandlingService,
+        @Autowired val databaseCleanupService: DatabaseCleanupService,
+        @Autowired val featureToggleService: FeatureToggleService,
+        @Autowired val integrasjonClient: IntegrasjonClient,
+        @Autowired val vedtakService: VedtakService,
 ) {
 
-    val søkerFnr = randomFnr()
+    val morsIdent = "04086226621"
+    val barnasIdenter = listOf("21111777001")
 
-    @BeforeAll
+    val clientMocks = ClientMocks()
+
+    @BeforeEach
     fun init() {
         databaseCleanupService.truncate()
+        clearAllMocks()
+        taskRepository.deleteAll()
+        initEuKodeverk(integrasjonClient)
+        mockPersonopplysning(morsIdent, mockSøkerAutomatiskBehandling, personopplysningerService)
+        mockIntegrasjonsClient(morsIdent, integrasjonClient)
+
+        every { featureToggleService.isEnabled(FeatureToggleConfig.AUTOMATISK_FØDSELSHENDELSE) } returns true
+        every { featureToggleService.isEnabled(FeatureToggleConfig.BRUK_VEDTAKSTYPE_MED_BEGRUNNELSER) } returns true
+
     }
 
     @Test
-    fun `sjekk om mor har løpende utbetalinger i infotrygd`() {
-        every { infotrygdService.harLøpendeSakIInfotrygd(mutableListOf(søkerFnr)) } returns true
+    fun `Søker uten løpende fagsak i BA blir sendt til infotrygd`() {
+        mockPersonopplysning(barnasIdenter.first(), mockBarnAutomatiskBehandling, personopplysningerService)
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        assertEquals(null, fagsak?.status)
 
-        assertEquals(true, velgFagSystemService.morHarLøpendeSakIInfotrygd(søkerFnr))
+        val taskFraInfotrygd = taskRepository.findAll()[0]
+        val InfotrygdFødselhendelsesFeedTaskDto =
+                objectMapper.readValue(taskFraInfotrygd.payload, InfotrygdFødselhendelsesFeedTaskDto::class.java)
+        assertEquals(barnasIdenter, InfotrygdFødselhendelsesFeedTaskDto.fnrBarn)
     }
 
     @Test
-    fun `skal velge Infotrygd når mor ikke har løpende sak i BA sak`() {
+    fun `Søker med løpende fagsak og betaling i BA blir sendt til manuell behandling`() {
+        mockPersonopplysning(barnasIdenter.first(), mockBarnAutomatiskBehandling, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(false)
 
-        fagSakService.hentEllerOpprettFagsak(PersonIdent(søkerFnr), true)
-        val nyBehandling = NyBehandlingHendelse(søkerFnr, listOf(søkerFnr))
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+        assertEquals(FagsakStatus.LØPENDE, åpenFagsak?.status)
 
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
 
-        assertEquals(velgFagSystemService.velgFagsystem(nyBehandling), FagsystemRegelVurdering.SEND_TIL_INFOTRYGD)
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        assertEquals(åpenFagsak.id, fagsak?.id)
+
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        //begynner neste behandling
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+        val fagsakEtterToBehandlinger = fagSakService.hent(PersonIdent(morsIdent))
+
+        val tasker = taskRepository.findAll()
+        val taskForOpprettelseAvManuellBehandling = tasker[1]
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals("Fødselshendelse: Bruker har åpen behandling", opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    fun `søker med løpende fagsak og verge blir sendt til manuell behandling `() {
+        val nyfødtBarn = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.now())
+        mockPersonopplysning(barnasIdenter.first(), nyfødtBarn, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(true)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(åpenFagsak.id, fagsak.id)
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        val taskForOpprettelseAvManuellBehandling = taskRepository.findAll().first()
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals(FiltreringsreglerResultat.MOR_HAR_VERGE.beskrivelse, opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    fun `søker med ugyldig Fnr blir sendt til manuell behandling`() {
+        val nyfødtBarn = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.now())
+        mockPersonopplysning(barnasIdenter.first(), nyfødtBarn, personopplysningerService)
+        val morsUgyldigeFnr = "04886226622"
+        mockIntegrasjonsClient(morsUgyldigeFnr, integrasjonClient)
+        mockPersonopplysning(morsUgyldigeFnr, mockSøkerAutomatiskBehandling, personopplysningerService)
+        every { personopplysningerService.harVerge(morsUgyldigeFnr) } returns VergeResponse(false)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsUgyldigeFnr), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsUgyldigeFnr, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsUgyldigeFnr))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(åpenFagsak.id, fagsak.id)
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        val taskForOpprettelseAvManuellBehandling = taskRepository.findAll().first()
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals(FiltreringsreglerResultat.MOR_IKKE_GYLDIG_FNR.beskrivelse, opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    fun `død søker blir sendt til manuell behandling`() {
+        val nyfødtBarn = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.now())
+        mockPersonopplysning(barnasIdenter.first(), nyfødtBarn, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(false)
+        every { personopplysningerService.hentDødsfall(Ident(morsIdent)) } returns DødsfallData(true, null)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(åpenFagsak.id, fagsak.id)
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        val taskForOpprettelseAvManuellBehandling = taskRepository.findAll().first()
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals(FiltreringsreglerResultat.MOR_ER_DØD.beskrivelse, opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    fun `søker med dødt barn blir sendt til manuell behandling`() {
+        val nyfødtBarn = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.now())
+        mockPersonopplysning(barnasIdenter.first(), nyfødtBarn, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(false)
+        every { personopplysningerService.hentDødsfall(Ident(barnasIdenter.first())) } returns DødsfallData(true, null)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(åpenFagsak.id, fagsak.id)
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        val taskForOpprettelseAvManuellBehandling = taskRepository.findAll().first()
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals(FiltreringsreglerResultat.DØDT_BARN.beskrivelse, opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    fun `søker under 18 blir sendt til manuell behandling`() {
+        val nyfødtBarn = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.now())
+        val morUnder18 = mockSøkerAutomatiskBehandling.copy(fødselsdato = LocalDate.now().minusYears(17))
+        mockPersonopplysning(morsIdent, morUnder18, personopplysningerService)
+        mockPersonopplysning(barnasIdenter.first(), nyfødtBarn, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(false)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(åpenFagsak.id, fagsak.id)
+        behandlingOgFagsakErÅpen(behanding, fagsak)
+
+        val taskForOpprettelseAvManuellBehandling = taskRepository.findAll().first()
+        val opprettOppgaveTaskDTO =
+                objectMapper.readValue(taskForOpprettelseAvManuellBehandling.payload, OpprettOppgaveTaskDTO::class.java)
+        assertEquals(behanding.id, opprettOppgaveTaskDTO.behandlingId)
+        assertEquals(FiltreringsreglerResultat.MOR_ER_IKKE_OVER_18.beskrivelse, opprettOppgaveTaskDTO.beskrivelse)
+    }
+
+    @Test
+    @Disabled
+    fun `Søker med barn over 18 går gjennom filtrering, stopper i vilkår`() {
+        val barnOver18 = mockBarnAutomatiskBehandling.copy(fødselsdato = LocalDate.parse("1999-10-10"))
+        mockPersonopplysning(barnasIdenter.first(), barnOver18, personopplysningerService)
+        every { personopplysningerService.harVerge(morsIdent) } returns VergeResponse(false)
+
+        val åpenFagsak = fagSakService.hentEllerOpprettFagsak(PersonIdent(morsIdent), true)
+        fagSakService.oppdaterStatus(åpenFagsak, FagsakStatus.LØPENDE)
+
+        lagOgkjørfødselshendelseTask(morsIdent, barnasIdenter, behandleFødselshendelseTask)
+
+        val fagsak = fagSakService.hent(PersonIdent(morsIdent))
+        val behanding = behandlingService.hentBehandlinger(fagsak!!.id).first()
+        assertEquals(BehandlingResultat.AVSLÅTT, behanding.resultat)
+        assertEquals(BehandlingÅrsak.FØDSELSHENDELSE, behanding.opprettetÅrsak)
+        assertEquals(StegType.VILKÅRSVURDERING, behanding.steg)
+        assertEquals(FagsakStatus.LØPENDE, fagsak.status)
+        assertEquals(åpenFagsak.id, fagsak.id)
     }
 }
+
